先初步的尝试一下，现在还有一些问题。

+++ Project1/Include/ResourceManager/ModelManager.h
@@ -13,7 +13,7 @@ private:
 	std::unordered_map<uint32_t, std::vector<uint32_t>> ModelBindMeshMap;
 
 	uint32_t modelID_ = 0;
-	size_t Basic_DynamicUniformBufferOffset = 0;
+	uint32_t Basic_DynamicUniformBufferOffset = 0;
 
 public:
 	//ModelManager();
@@ -40,7 +40,7 @@ public:
 	std::shared_ptr<modelInstance>& getModelInstanceByID(uint32_t modelID) { return bindModels[modelID]; }
 	std::unordered_map<uint32_t, std::shared_ptr<modelInstance>>& getModelInstances() { return bindModels; }
 
-	void setOffest(size_t offset) { Basic_DynamicUniformBufferOffset = offset; }
+	void setOffest(uint32_t offset) { Basic_DynamicUniformBufferOffset = offset; }
 	uint32_t getOffeset() { return Basic_DynamicUniformBufferOffset; }
